refactor(drawer): centralize drawer navigation logic
Extracts common logic for building drawer links and closing drawers into a shared utility. This reduces code duplication across dashboard, import, and summary drawer navigation modules, promoting consistency and maintainability.
