+++ projects/client/src/lib/components/drawer/constants/index.ts
@@ -1,0 +1,1 @@
+export const DRAWER_VIEW_PARAM = 'view';
