+++ projects/client/src/lib/sections/dashboard/_internal/dashboardDrawerNavigation.ts
@@ -1,7 +1,5 @@
-import { goto } from '$app/navigation';
-import { page } from '$app/state';
-
-const VIEW_PARAM = 'view';
+import { DRAWER_VIEW_PARAM } from '$lib/components/drawer/constants/index.ts';
+import { drawerNavigation } from '$lib/components/drawer/drawerNavigation.ts';
 
 export enum DashboardDrawers {
   WeeklyPulse = 'weekly-pulse',
@@ -22,23 +20,10 @@ function mapToDrawer(value: string | Nil) {
 export function dashboardDrawerNavigation(
   searchParams?: URLSearchParams,
 ) {
-  const drawer = mapToDrawer(searchParams?.get(VIEW_PARAM));
-
-  const buildDrawerLink = (drawer: DashboardDrawers) => {
-    const url = new URL(page.url);
-    url.searchParams.set(VIEW_PARAM, drawer);
-    return url.toString();
-  };
-
-  const close = () => {
-    const url = new URL(page.url);
-    url.searchParams.delete(VIEW_PARAM);
-    goto(url, { noScroll: true });
-  };
+  const drawer = mapToDrawer(searchParams?.get(DRAWER_VIEW_PARAM));
 
   return {
     drawer,
-    buildDrawerLink,
-    close,
+    ...drawerNavigation<DashboardDrawers>(),
   };
 }
