+++ projects/client/src/lib/sections/summary/_internal/summaryDrawerNavigation.ts
@@ -1,7 +1,5 @@
-import { goto } from '$app/navigation';
-import { page } from '$app/state';
-
-const VIEW_PARAM = 'view';
+import { DRAWER_VIEW_PARAM } from '$lib/components/drawer/constants/index.ts';
+import { drawerNavigation } from '$lib/components/drawer/drawerNavigation.ts';
 
 export enum Drawers {
   Sentiment = 'sentiment',
@@ -38,23 +36,10 @@ function mapToDrawer(value: string | Nil) {
 }
 
 export function summaryDrawerNavigation(searchParams?: URLSearchParams) {
-  const drawer = mapToDrawer(searchParams?.get(VIEW_PARAM));
-
-  const buildDrawerLink = (drawer: Drawers) => {
-    const url = new URL(page.url);
-    url.searchParams.set(VIEW_PARAM, drawer);
-    return url.toString();
-  };
-
-  const close = () => {
-    const url = new URL(page.url);
-    url.searchParams.delete(VIEW_PARAM);
-    goto(url, { noScroll: true });
-  };
+  const drawer = mapToDrawer(searchParams?.get(DRAWER_VIEW_PARAM));
 
   return {
     drawer,
-    buildDrawerLink,
-    close,
+    ...drawerNavigation<Drawers>(),
   };
 }
